\name{gsg-package}
\alias{gsg-package}
\alias{gsg}
\docType{package}
\title{
Calculation of selection coefficients
}
\description{
Functions to calculate selection gradients, especially \code{gam.gradients()} which allows selection gradients to be calcuated from a generalized additive model-based characterization of a fitness function.
}
\details{
\tabular{ll}{
Package: \tab gsg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-23\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}

\author{
Michael Morrissey

Maintainer: Michael Morrissey \email{michael.morrissey@st-andrews.ac.uk}
}
\references{
M.B. Morrissey and K. Sakrejda. 2013. Unification of regression-based methods for the analysis of natural selection. Evolution 67: 2094-2100.

M.B. Morrissey. in prep.
}

\keyword{ package }
