\name{humanNeonatal}
\docType{data}
\alias{humanNeonatal}
\title{Human male birth mass, gestation length, and neonatal survival}
\description{
  Karn and Penrose's (1951) dataset on the survival of human male infants, their birth wieghts in 0.23 kg (half pound) increments, and their gestation lengths in 5 day increments. 
}
\source{Transcribed by Dolph Schluter from  Karn and Penrose 1951.}
\references{
Karn, M. N., and L. S. Penrose. 1951. Birth weight and gestation time in relation to maternal age, parity and infant survival.  Annals fo Eugenics 16: 147-164.

Schluter, D. 1988. Estimating the form of natural selection on a quantitative trait. Evolution 42: 849-861.

Schluter, D., and D. Nychka. 1994. Exploring fitness surfaces.  The american Naturalist 143: 597-616.

M.B. Morrissey and K. Sakrejda. 2013. Unification of regression-based methods for the analysis of natural selection. Evolution 67: 2094-2100.
}

\usage{humansNeonatal}
\format{a data frame}
\keyword{datasets}

