% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_design}
\alias{plot_design}
\title{Add point clusters at the points of the Global Sampling Grid}
\usage{
plot_design(gsg, layout, ref = dim(layout)/2 + 0.5, dist, rot = 0)
}
\arguments{
\item{gsg}{A matrix with longitudes (first column) and latitudes (second column).}

\item{layout}{A matrix specifiying the layout of the point cluster. See Details and Examples.}

\item{ref}{Where is the reference of the point cluster. With default settings
it is centered on the grid coordinates. Defaults can be overwritten by
providing row and column indexes. See examples.}

\item{dist}{Distance in m between points in the cluster.}

\item{rot}{Should the clusters be rotated? Default is 0 for no rotation. If
'random', each point cluster is rotated in a random direction. If a single
value is provided, all clusters are rotated in the same direction.}
}
\value{
SpatialPointsDataFrame. The data slot contains point and sub-point IDs.
}
\description{
Add point clusters at the points of the Global Sampling Grid
}
\examples{
library(gsg);
# Generate a global sampling grid with 500km distance
gsg <- gen_gsg(500);

# Define the layout of the point cluster
# 25 points arranged in a square
layout <- matrix(rep(1, 25), nrow = 1);
# 500m between points
dist = 500;
# Centered on the grid points
ref <- c(3, 3);

# New grid with clusters
gsg_clus <- plot_design(gsg = gsg@coords,
                        layout = layout,
                        ref = ref,
                        dist = dist,
                        rot = 0);
plot(gsg_clus)

# L-shaped clusters
layout <- matrix(c(1, 0, 0,
                   1, 0, 0,
                   1, 0, 0,
                   1, 1, 1), nrow = 4, byrow = TRUE);

# Cross
layout <- matrix(c(0, 1, 0,
                   1, 1, 1,
                   0, 1, 1), nrow = 3);

# Export to Shapefile
rgdal::writeOGR(gsg_cl, "gsg_clus.shp", driver = "ESRI Shapefile", layer = "gsg_clus");
}
