% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_server.R
\name{shiny_app_server}
\alias{shiny_app_server}
\title{Shiny app server function}
\usage{
shiny_app_server(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}
}
\description{
Shiny app server function
}
\examples{
# Start shiny app using launch_app() function
launch_app()
}
