% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_ui.R
\docType{data}
\name{shiny_app_ui}
\alias{shiny_app_ui}
\title{Shiny app server object}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
shiny_app_ui
}
\description{
Shiny app server object
}
\keyword{datasets}
