\name{SoayLambs}
\docType{data}
\alias{SoayLambs}
\title{Soay sheep lamb phenotypic data and first year winter survival}
\description{
  Data for mass, hind leg length, horn length, and (log) ked count of Soay sheep measured in August.  Fitness (W) is first-winter survival.  Data are subsetted to females with normal horns in cohorts that did not experience crash conditions in their first winter.  Phenotypic data are mean-centred and variance-standardized.
}
\source{Data provided by Profs. J.M. Pemberton and L.E.B. Kruuk, University of Edinburgh.  The data were collected primarily by Jill Pemberton and Andrew MacColl with the help of many volunteers, with logistical support from the National Trust for Scotland and QinetQ, and with funding from NERC, Royal Society, and the Leverhulme Trust.}
\references{
Clutton-Brock, T.H., and J.M. Pemberton. 2004. Soay sheep dynamics an dselection in an island population. Cambridge University Press, Cambridge.

Morrissey, M. B. in prep.
}

\usage{SoayLambs}
\format{a data frame}
\keyword{datasets}

